Playing with regular expressions
I had a case where I was using a couple of regular expressions to parse out tokens from a line but when I was starting to use different lines, the regular expressions were not working.  Rather than come up with two regular expressions for each type of message, I modified the regular expressions I had to work with both types of lines.  I made use of the `(?:...)` construct to avoid capturing portions of the line so the captured substrings remain the same.