Fixing problem with passlogger
The problem was that my other script was dealing with unicode strings that were loaded from a JSON document.  My regular expression identified the `u` prefix as the password rather than realizing that it's just Python's way to saying "the string is unicode".  This change teaches the script how to know when there's a prefix and how to skip it.