Solved the popen/pipe problem!
My problem was that I was both ends of the pipe (read & write) were open in the child process!  Since the write end was not closed, the child would keep trying to read from the other end!  Using the `close_fds` parameter of the subprocess.Popen() to close the write end of the pipe in the child worked.