Initial version of alterate Table class
I've had a Table class for a while that was kind of complex.  I was hacking about with a sort of CLI and developed what may be a more simple Table class.

One thing I want to do with this is remove the dependency on the subprocess class.  That will also make it so that I can pad numeric columns on the left.