Simplifying widths/numeric logic
I was calculating the widths and numeric states of columns each time the __str__() method was executed but it seemed to make more send to store these as instance variables and maintain them as append() is called.  This also simplifies the logic when headings are specified.  The append() methods doesn't even care if headings are specified - it only deals with data in the table.