Starting switch from Python 2 to Python 3
I'm working on changing scripts in my toys directory to Python 3.  I want this fragment to call out for Python 3 from now on.  I haven't quite changed the toys directory yet but I want this changed first.  If I have the need to use the fragment without using Python 3, I can easily change the new script that uses the fragment on a case-to-case basis.  But I don't think there will be a need to do that!  I should be able to just continue with Python 3.