Adding SIGPIPE handler
I just made this change to git-log and I want to add it to the `basic` script on which I use to model new scripts to have the capability by default.  The handler will mask SIGPIPE errors when writing to stdout as a pipe after the pipe has been closed.