Commenting out import for sys module
I was including the file in a new script and noticed it was including the sys module but not using it.  I was tempted to just delete it but for some reason I thought there was a reason I did this... like there was a good reason it was there.  Anyway, I commented it out.